Document and split off token tree and location structures
